Важный коммит! Реализована автоматическая генерация второго девелоп окна. Пока автоматизация заключается в количестве полей для изменения характеристик -- их будет столько, сколько указали в предыдущем окне. Продолжаются работы

+++ GameConstructor.Core/Models/Effect.cs
@@ -9,9 +9,14 @@ namespace GameConstructor.Core.Models
     public class Effect
     {
         public int Id { get; set; }
+
         public string Body { get; set; }
+
         public List<Influence> Influences { get; set; }
-        public double Value { get; set; }
+
+        public int Value { get; set; }
+
+
         public void Influenced()
         {
             foreach ( var influence in Influences)
